backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.7.0
pydantic-settings==2.2.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
PyJWT==2.8.1
python-multipart==0.0.6
bcrypt==4.1.2
cryptography==42.0.2
python-dotenv==1.0.0
email-validator==2.1.0
passlib==1.7.4
